+++ pygpsclient/_version.py
@@ -8,4 +8,4 @@ Created on 12 Sep 2020
 :license: BSD 3-Clause
 """
 
-__version__ = "1.0.4"
+__version__ = "1.0.5"
